+++ app/src/main/java/cn/ianzb/miuixguitemplate/TemplateApp.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import cn.ianzb.miuixguitemplate.prefs.ConfigState
 import cn.ianzb.miuixguitemplate.prefs.OptionRegistry
 import cn.ianzb.miuixguitemplate.prefs.PrefsStore
-import cn.ianzb.miuixguitemplate.ui.screen.examples.exampleSpecs
+import cn.ianzb.miuixguitemplate.ui.screen.features.featureSpecs
 import cn.ianzb.miuixguitemplate.xposed.XposedServiceManager
 
 class TemplateApp : Application() {
@@ -13,7 +13,7 @@ class TemplateApp : Application() {
         super.onCreate()
         PrefsStore.init(this)
         ConfigState.init(this)
-        OptionRegistry.registerAll(exampleSpecs())
+        OptionRegistry.registerAll(featureSpecs())
         XposedServiceManager.init()
     }
 }